+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Improve compiler error message for missing Swift declarations due to APPLICATION_EXTENSION_API_ONLY (#4603)
 
+### Fixes
+
+- `SentrySdkInfo.packages` should be an array (#4626)
+
 ## 8.42.0-beta.2
 
 ### Fixes